LayoutTests/imported/w3c/web-platform-tests/navigation-api/ordering-and-transition/currententrychange-dispose-ordering-expected.txt


FAIL Ordering between Navigation currententrychange and NavigationHistoryEntry dispose events assert_equals: expected -1 but got 1
